requirements.txt
cachetools>=5.5.0
icalendar>=6.0.0
pytz>=2022.1
pyyaml>=5.4.1
recurring-ical-events
requests>=2.28.1

Flask>=3.0.3